Initialize FuseNet Object

Initialize a FuseNet object.

InitiateFuseNet(
  raw_data,
  project_name = "",
  normalization = c("cosine", "lognorm", "none"),
  normalize_factor = 10000,
  zero_percent = 0.7,
  pca_dims = 0,
  kernel = c("gaussian", "euclidean"),
  k = 100,
  t = 0,
  verbose = FALSE,
  seed = 1
)

Arguments

raw_data

Raw data. An N x M matrix with N rows of features and M columns of data points.

project_name

Project name. Default is none.

normalization

Normalization method used. Default is cosine. See details Normalization.

cosine, cosine normalization: feature counts for each data point are divided by the L2 norm of them.

lognorm, log normalization: feature counts for each data point are divided by the total sum of them. Then the data is multiplied by the scale.factor before taking a log-transformed by log(1+x).

none, additional normalization is not performed.

normalize_factor

Normalization factor used with lognorm method. Default is 10000.

zero_percent

Zero-entry percentage threshold. If the number of zero entries in the returned matrices is above this number, a sparse matrix will be returned. Default is 0.7 aka 70%.

pca_dims

Number of dimensions used. Default is 0 and PCA is not performed.

kernel

Kernel distance used:

gaussian, gaussian distance kernel. See details EuclideanDist.

euclidean, euclidean distance kernel. See details GaussianDist.

k

Number of nearest neighbors. Default is 100. See details from nn2.

t

Matrix power used for the distance matrix. Default is 0 and powering is not performed. See MatrixPower for details.

verbose

Whether to display a process bar. Default is FALSE.

seed

Random seed number. Default is 1.

Value

Returns a FuseNet object.

Examples

{
object <- InitiateFuseNet(t(iris[,1:4]), project_name = "FuseNet", k = 10)
}
#> Initiate FuseNet
#> Normalize Data
#> Find Nearest Neighbors
#> Matrix Power
#> Finalize